requirements.txt
openai
pythondotenv
langsmith